refactor: move Beatmaps tab to separate component
